+++ spec/rubocop/cop/style/if_with_semicolon_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe RuboCop::Cop::Style::IfWithSemicolon do
   it 'registers an offense and corrects for one line if/;/end' do
     expect_offense(<<~RUBY)
       if cond; run else dont end
-      ^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not use if x; Use the ternary operator instead.
+      ^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not use `if cond;` - use a ternary operator instead.
     RUBY
 
     expect_correction(<<~RUBY)
@@ -27,5 +27,58 @@ RSpec.describe RuboCop::Cop::Style::IfWithSemicolon do
       class Hash
       end if RUBY_VERSION < "1.8.7"
     RUBY
+  end
+
+  context 'when elsif is present' do
+    it 'accepts without `else` branch' do
+      expect_offense(<<~RUBY)
+        if cond; run elsif cond2; run2 end
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not use `if cond;` - use `if/else` instead.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        if cond
+          run
+        elsif cond2
+          run2
+        end
+      RUBY
+    end
+
+    it 'accepts second elsif block' do
+      expect_offense(<<~RUBY)
+        if cond; run elsif cond2; run2 elsif cond3; run3 else dont end
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not use `if cond;` - use `if/else` instead.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        if cond
+          run
+        elsif cond2
+          run2
+        elsif cond3
+          run3
+        else
+          dont
+        end
+      RUBY
+    end
+
+    it 'accepts with `else` branch' do
+      expect_offense(<<~RUBY)
+        if cond; run elsif cond2; run2 else dont end
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not use `if cond;` - use `if/else` instead.
+      RUBY
+
+      expect_correction(<<~RUBY)
+        if cond
+          run
+        elsif cond2
+          run2
+        else
+          dont
+        end
+      RUBY
+    end
   end
 end